eeshow/gui/input.c (clicked_x, _y): avoid accumulation of conversion errors
Gtk uses floating-point coordinates. The double conversion during dragging
(event -> clicked -> difference with next event) caused dragging to "slow
down" in the X- and Y- direction.

+++ gui/input.c
@@ -42,7 +42,7 @@ static struct input {
 } *sp = NULL;
 
 static int curr_x, curr_y;		/* last mouse position */
-static int clicked_x, clicked_y;	/* button down position */
+static double clicked_x, clicked_y;	/* button down position */
 
 
 /* ----- Debugging tools --------------------------------------------------- */
